// packages/api-client/models/NotificationCreateRo.ts
/**
 * Api Document
 * Backend_Server Api Document
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { JSONObject } from '../models/JSONObject';
import { HttpFile } from '../http/http';

/**
* User notification list parameters
*/
export class NotificationCreateRo {
    /**
    * ID of the notified user (optional)
    */
    'toUserId'?: Array<string>;
    /**
    * Either the member ID or to User Id of the notified user (optional)
    */
    'toMemberId'?: Array<string>;
    /**
    * Either the organizational unit ID or to User ID of the notified user (optional)
    */
    'toUnitId'?: Array<string>;
    /**
    * Send the notification user ID, and the system notifies the user as 0 (optional)
    */
    'fromUserId'?: string;
    /**
    * Node ID (optional)
    */
    'nodeId'?: string;
    /**
    * Space ID (optional)
    */
    'spaceId'?: string;
    /**
    * Template ID
    */
    'templateId': string;
    'body'?: JSONObject;
    /**
    * Version number (optional)
    */
    'version'?: string;
    /**
    * Expiration time (optional) accurate to milliseconds
    */
    'expireAt'?: string;
    /**
    * Notification ID (optional)
    */
    'notifyId'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "toUserId",
            "baseName": "toUserId",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "toMemberId",
            "baseName": "toMemberId",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "toUnitId",
            "baseName": "toUnitId",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "fromUserId",
            "baseName": "fromUserId",
            "type": "string",
            "format": ""
        },
        {
            "name": "nodeId",
            "baseName": "nodeId",
            "type": "string",
            "format": ""
        },
        {
            "name": "spaceId",
            "baseName": "spaceId",
            "type": "string",
            "format": ""
        },
        {
            "name": "templateId",
            "baseName": "templateId",
            "type": "string",
            "format": ""
        },
        {
            "name": "body",
            "baseName": "body",
            "type": "JSONObject",
            "format": ""
        },
        {
            "name": "version",
            "baseName": "version",
            "type": "string",
            "format": ""
        },
        {
            "name": "expireAt",
            "baseName": "expireAt",
            "type": "string",
            "format": ""
        },
        {
            "name": "notifyId",
            "baseName": "notifyId",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return NotificationCreateRo.attributeTypeMap;
    }

    public constructor() {
    }
}